move implementation of functions to cpp file
